Move resultify into its own module
