+++ packages/bbrt/src/breadboard/indexed-db-servers.ts
@@ -4,7 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { resultify, type Result } from "../util/result.js";
+import type { Result } from "../util/result.js";
+import { resultify } from "../util/resultify.js";
 import { BreadboardServer } from "./breadboard-server.js";
 
 type BoardServerEntry = { url: string };
